alsa: make corrections on the timeout based on how fast ALSA consumes samples
This feels a bit hacky, but it actually makes huge difference when pipewire is
running in qemu.

The idea is that it keeps track of how much samples are in the device
(fill level) and calculates how many are consumed when a timeout occurs.
Then it converts that into a time based on the sample rate and compares it to
the system clock time that elapsed since the last write to the device.
The division between the two gives a rate (drift) that can be used to shorten
the timeout window.

So for instance, if the timeout window was 21.3 ms, but the device actually
consumed an equivalent of 28 ms in samples, the drift will be 21.3/28 = 0.76
and the next timeout window will be approximately 21.3 * 0.76 = 16.1 ms

To avoid making things worse, the drift is clamped between 0.6 and 1.0.
Min 0.6 was arbitrarily chosen, but sometimes alsa does report strange numbers,
causing the drift to be very low, which in turn causes an early wakeup.
Max 1.0 basically means that we don't care if the device is consuming samples
slower. In that case, the early wakeup mechanism will throttle pipewire.

Fixes #163