gst/pwaudioringbuffer: wait only for STREAM_STATE_CONFIGURE when starting
The CONFIGURE state is reached when the pw_client_node is exported,
while the READY state requires the session manager to try and link
the stream. If the SM does not want to link the stream due to policy,
the client should not hang there forever.

Upstream-Status: Submitted [https://github.com/PipeWire/pipewire/pull/140]